+++ ubcore.sh
@@ -2298,7 +2298,8 @@ _discoverResource() {
 _rmlink() {
 	[[ "$1" == "/dev/null" ]] && return 1
 	
-	[[ -h "$1" ]] && rm -f "$1" && return 0
+	#[[ -h "$1" ]] && rm -f "$1" && return 0
+	[[ -h "$1" ]] && rm -f "$1" > /dev/null 2>&1
 	
 	! [[ -e "$1" ]] && return 0
 	
@@ -9934,11 +9935,15 @@ _close_sequence() {
 	
 	if [[ "$?" == "0" ]]
 	then
-		rm -f "$lock_open" || return 1
+		#rm -f "$lock_open" || return 1
+		rm -f "$lock_open"
+		[[ -e "$lock_open" ]] && return 1
 		
 		if [[ "$specialLock" != "" ]] && [[ -e "$specialLock" ]]
 		then
-			rm -f "$specialLock" || return 1
+			#rm -f "$specialLock" || return 1
+			rm -f "$specialLock"
+			[[ -e "$specialLock" ]] && return 1
 		fi
 		
 		rm -f "$lock_closing"
@@ -10844,6 +10849,9 @@ _test_embed() {
 }
 
 _test_sanity() {
+	# Do NOT allow 'rm' to be a shell function alias to 'rm -i' or similar.
+	[[ $(type -p rm) == "" ]] && _messageFAIL && return 1
+	
 	#! [[ -2147483648 -lt 2147483647 ]] && _messageFAIL && return 1
 	#! [[ -2000000000 -lt 2000000000 ]] && _messageFAIL && return 1
 	
@@ -10899,6 +10907,7 @@ _test_sanity() {
 	
 	[[ ! -e "$safeTmp" ]] && _messageFAIL && return 1
 	
+	
 	local currentTestUID=$(_uid 245)
 	mkdir -p "$safeTmp"/"$currentTestUID"
 	echo > "$safeTmp"/"$currentTestUID"/"$currentTestUID"
@@ -10924,6 +10933,16 @@ _test_sanity() {
 	_define_function_test
 	
 	! _variableLocalTest && _messageFAIL && return 1
+	
+	
+	
+	mkdir -p "$safeTmp"/maydeletethisfolder
+	[[ ! -d "$safeTmp"/maydeletethisfolder ]] && return 1
+	echo > "$safeTmp"/maydeletethisfolder/maydeletethisfile
+	[[ ! -e "$safeTmp"/maydeletethisfolder/maydeletethisfile ]] && return 1
+	_safeRMR "$safeTmp"/maydeletethisfolder
+	[[ -e "$safeTmp"/maydeletethisfolder/maydeletethisfile ]] && return 1
+	[[ -e "$safeTmp"/maydeletethisfolder ]] && return 1
